+++ mini_agent/session.py
@@ -3,8 +3,6 @@ class Session:
         self.session_id = session_id
         self.memory = memory
         self.memory.create_session(self.session_id)
-        self.messages = self.memory.get_messages(self.session_id)
-
 
     def add_user_message(self, content):
         message = {
@@ -12,22 +10,6 @@ class Session:
             "content": content
             }
         
-        self.messages.append(message)
-
-        self.memory.add_message(
-            self.session_id,
-            message
-        )
-
-
-    def add_system_message(self, content):
-        message = {
-            "role": "system", 
-            "content": content
-            }
-        
-        self.messages.append(message)
-
         self.memory.add_message(
             self.session_id,
             message
@@ -35,15 +17,13 @@ class Session:
 
 
     def add_tool_message(self, tool_call_id, tool_name, content):
-        
+
         message = {
             "role": "tool",
             "tool_call_id": tool_call_id,
-            "tool_name": tool_name,
+            "name": tool_name,
             "content": str(content)
         }
-
-        self.messages.append(message)
 
         self.memory.add_message(
             self.session_id,
@@ -68,16 +48,8 @@ class Session:
                     }
                 })
 
-        self.messages.append(data)
-
         self.memory.add_message(
             self.session_id,
             data
         )
 
-
-    def get_messages(self):
-        return self.messages
-
-    def get_recent_messages(self, n=5):
-        return self.messages[-n:]